testify/requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.5
flask-login==0.6.2
dash==2.9.3
dash-bootstrap-components==1.4.1
pandas==2.0.3
openpyxl==3.1.2
pymysql==1.1.0
weasyprint==59.0
python-dotenv==1.0.0